+++ WEB/BE/controllers/auth/index.js
@@ -3,7 +3,7 @@ const authService = require('@services/auth');
 const userService = require('@services/user');
 const { comparePassword, getEncryptedPassword } = require('@utils/bcrypt');
 const { encryptWithAES256, decryptWithAES256 } = require('@utils/crypto');
-const { emailSender } = require('@utils/email');
+const { emailSender } = require('@/utils/emailSender');
 const createError = require('http-errors');
 const JWT = require('jsonwebtoken');
